requirements.txt
aiohttp
telethon
python-dotenv